(refactor) Inline hosts
